refactor(jsx): decompose 3 sibling setup wizards (PR-10/11)
Applies the operator-setup-wizard pattern (PR-portal-4 / #206) to
the 3 remaining setup wizards. Each gets a sibling subdirectory
with fixtures (data tables) + utils (generator functions) extracted;
Step components remain inline for a future PR.

Per-wizard shrink
-----------------
* cicd-setup-wizard.jsx:    642 → 504 LOC (-22%)
* deployment-wizard.jsx:    925 → 504 LOC (-46%)
* rbac-setup-wizard.jsx:    708 → 648 LOC (-9%)
* TOTAL:                  2275 → 1656 LOC (-619, -27%)

All 3 stay under the sed-damage-guard 50% threshold — no .sed-damage-
allowlist entry needed (only PR-9's portal-shared.jsx required it
because of the 62% shim shrink).

New files (6)
-------------
* cicd-setup-wizard/fixtures/wizard-defaults.js — STEPS, RULE_PACKS,
  CI_OPTIONS, DEPLOY_OPTIONS (~70 LOC)
* cicd-setup-wizard/utils/generators.js — generateInitCommand,
  generateDockerCommand, generateFileTree,
  generateGitHubActionsPreview (~140 LOC)
* deployment-wizard/fixtures/wizard-defaults.js — STEPS, TIERS,
  ENVIRONMENTS, TENANT_SIZES, OAUTH2_PROVIDERS, RULE_PACKS (~165 LOC)
* deployment-wizard/utils/generators.js — generateHelmValues
  (~310 LOC YAML template)
* rbac-setup-wizard/fixtures/wizard-defaults.js — STEPS,
  PERMISSION_HIERARCHY, ENVIRONMENTS, DOMAIN_EXAMPLES (~50 LOC)
* rbac-setup-wizard/utils/generators.js — generateRbacYaml,
  validateRbac (~80 LOC)

Symbol-naming convention to avoid window-global collisions across
wizards (each wizard prefixes its symbols):
  __CICD_*, __cicdGenerate*       (cicd-setup-wizard)
  __DEPLOY_*, __deployGenerate*   (deployment-wizard)
  __RBAC_*, __rbacGenerate*       (rbac-setup-wizard)

Each orchestrator updates its front-matter dependencies block to
load the 2 sibling files, then picks up symbols off window. No
behaviour change — the extracted code is byte-for-byte identical
(verified by diff against HEAD).

Verification
------------
* All 38 pre-commit hooks green
* All 6 new dep paths resolve on disk
* lint_jsx_babel --ci --strict-linecount exit 0; static warnings
  unchanged (350) — extraction moved violations between files but
  total count identical
* tool-consistency-check clean
* No allowlist entries needed (each wizard stays > 50% of original)

Roadmap
-------
PR-10 of 11. PR-11 (subtree ErrorBoundaries on self-service-portal
3 tabs + multi-step wizards) is the closer.
