+++ docs/interactive/tools/cicd-setup-wizard.jsx
@@ -5,166 +5,28 @@ audience: ["platform-engineer"]
 version: v2.7.0
 lang: en
 related: [self-service-portal, template-gallery, onboarding-checklist]
+dependencies: [
+  "cicd-setup-wizard/fixtures/wizard-defaults.js",
+  "cicd-setup-wizard/utils/generators.js"
+]
 ---
 
 import React, { useState, useMemo } from 'react';
 
 const t = window.__t || ((zh, en) => en);
 
-/* ── Step definitions ── */
-const STEPS = [
-  { id: 'ci', label: () => t('CI/CD 平台', 'CI/CD Platform') },
-  { id: 'deploy', label: () => t('部署方式', 'Deployment Mode') },
-  { id: 'packs', label: () => t('Rule Packs', 'Rule Packs') },
-  { id: 'tenants', label: () => t('Tenant 設定', 'Tenant Setup') },
-  { id: 'review', label: () => t('檢視與產出', 'Review & Generate') },
-];
+// PR-portal-10: data + helpers extracted to sibling subdirectory
+// (mirrors operator-setup-wizard PR-portal-4 pattern). Step
+// components remain inline; future PR can extract them.
+const STEPS = window.__CICD_STEPS;
+const RULE_PACKS = window.__CICD_RULE_PACKS;
+const CI_OPTIONS = window.__CICD_CI_OPTIONS;
+const DEPLOY_OPTIONS = window.__CICD_DEPLOY_OPTIONS;
 
-/* ── Rule Pack catalog ── */
-const RULE_PACKS = [
-  { id: 'mariadb', label: 'MariaDB/MySQL', category: 'database', icon: '🐬', defaultOn: true },
-  { id: 'postgresql', label: 'PostgreSQL', category: 'database', icon: '🐘' },
-  { id: 'redis', label: 'Redis', category: 'database', icon: '🔴' },
-  { id: 'mongodb', label: 'MongoDB', category: 'database', icon: '🍃' },
-  { id: 'elasticsearch', label: 'Elasticsearch', category: 'database', icon: '🔎' },
-  { id: 'oracle', label: 'Oracle', category: 'database', icon: '🏛️' },
-  { id: 'db2', label: 'DB2', category: 'database', icon: '🔷' },
-  { id: 'clickhouse', label: 'ClickHouse', category: 'database', icon: '🖱️' },
-  { id: 'kafka', label: 'Kafka', category: 'messaging', icon: '📨' },
-  { id: 'rabbitmq', label: 'RabbitMQ', category: 'messaging', icon: '🐰' },
-  { id: 'jvm', label: 'JVM', category: 'runtime', icon: '☕' },
-  { id: 'nginx', label: 'Nginx', category: 'webserver', icon: '🌐' },
-  { id: 'kubernetes', label: 'Kubernetes', category: 'infrastructure', icon: '⎈', defaultOn: true },
-];
-
-const CI_OPTIONS = [
-  { id: 'github', label: 'GitHub Actions', icon: '🐙', desc: () => t('最廣泛使用的 CI/CD，直接整合 GitHub', 'Most widely used CI/CD, built into GitHub.') },
-  { id: 'gitlab', label: 'GitLab CI', icon: '🦊', desc: () => t('GitLab 內建 CI/CD，支援自託管', 'GitLab built-in CI/CD, supports self-hosted.') },
-  { id: 'both', label: t('兩者皆有', 'Both'), icon: '🔄', desc: () => t('同時產生 GitHub Actions 和 GitLab CI 配置', 'Generate both GitHub Actions and GitLab CI configs.') },
-];
-
-const DEPLOY_OPTIONS = [
-  { id: 'kustomize', label: 'Kustomize', icon: '📦', desc: () => t('推薦入門：configMapGenerator 自動產生 ConfigMap', 'Recommended to start: configMapGenerator auto-creates ConfigMap.') },
-  { id: 'helm', label: 'Helm', icon: '⛵', desc: () => t('使用 threshold-exporter Helm chart 管理', 'Managed via threshold-exporter Helm chart.') },
-  { id: 'argocd', label: 'ArgoCD', icon: '🔁', desc: () => t('GitOps 自動同步：ArgoCD Application 指向你的 repo', 'GitOps auto-sync: ArgoCD Application points to your repo.') },
-];
-
-/* ── Generate da-tools init command ── */
-function generateInitCommand(config) {
-  const parts = ['da-tools init'];
-  if (config.ci) parts.push(`--ci ${config.ci}`);
-  if (config.deploy) parts.push(`--deploy ${config.deploy}`);
-  if (config.tenants.length > 0) parts.push(`--tenants ${config.tenants.join(',')}`);
-  if (config.packs.length > 0) parts.push(`--rule-packs ${config.packs.join(',')}`);
-  parts.push('--non-interactive');
-  return parts.join(' \\\n  ');
-}
-
-/* ── Generate docker run command ── */
-function generateDockerCommand(config) {
-  const init = generateInitCommand(config);
-  return `docker run --rm -it \\\n  -v $(pwd):/workspace -w /workspace \\\n  ghcr.io/vencil/da-tools:latest \\\n  ${init.replace('da-tools ', '')}`;
-}
-
-/* ── Generate file tree preview ── */
-function generateFileTree(config) {
-  const lines = ['your-repo/'];
-  lines.push('├── conf.d/');
-  lines.push('│   ├── _defaults.yaml');
-  for (const tenant of config.tenants) {
-    lines.push(`│   └── ${tenant}.yaml`);
-  }
-  if (config.ci === 'github' || config.ci === 'both') {
-    lines.push('├── .github/workflows/');
-    lines.push('│   └── dynamic-alerting.yaml');
-  }
-  if (config.ci === 'gitlab' || config.ci === 'both') {
-    lines.push('├── .gitlab-ci.d/');
-    lines.push('│   └── dynamic-alerting.yml');
-  }
-  if (config.deploy === 'kustomize' || config.deploy === 'argocd') {
-    lines.push('├── kustomize/');
-    lines.push('│   ├── base/');
-    lines.push('│   │   └── kustomization.yaml');
-    lines.push('│   └── overlays/');
-    lines.push('│       ├── dev/');
-    lines.push('│       └── prod/');
-  }
-  if (config.deploy === 'argocd') {
-    lines.push('├── argocd/');
-    lines.push('│   └── dynamic-alerting.yaml');
-  }
-  lines.push('├── .pre-commit-config.da.yaml');
-  lines.push('└── .da-init.yaml');
-  return lines.join('\n');
-}
-
-/* ── Generate GitHub Actions YAML preview ── */
-function generateGitHubActionsPreview(config) {
-  return `name: Dynamic Alerting CI/CD
-on:
-  pull_request:
-    paths: ['conf.d/**']
-  push:
-    branches: [main]
-    paths: ['conf.d/**']
-  workflow_dispatch:
-
-jobs:
-  validate:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-      - name: Validate config
-        run: |
-          docker run --rm \\
-            -v \${{ github.workspace }}/conf.d:/data/conf.d:ro \\
-            ghcr.io/vencil/da-tools:latest \\
-            validate-config --config-dir /data/conf.d --ci
-
-  generate:
-    needs: validate
-    if: github.event_name == 'pull_request'
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-      - name: Prepare output directory
-        run: mkdir -p .output
-      - name: Generate routes
-        run: |
-          docker run --rm \\
-            -v \${{ github.workspace }}/conf.d:/data/conf.d:ro \\
-            -v \${{ github.workspace }}/.output:/data/output \\
-            ghcr.io/vencil/da-tools:latest \\
-            generate-routes --config-dir /data/conf.d -o /data/output/routes.yaml --validate
-      - name: Compute blast radius
-        run: |
-          docker run --rm \\
-            -v \${{ github.workspace }}/conf.d:/data/conf.d:ro \\
-            ghcr.io/vencil/da-tools:latest \\
-            config-diff --old-dir /data/conf.d.base --new-dir /data/conf.d --format markdown > .output/blast-radius.md
-
-  apply:
-    needs: [validate, generate]
-    if: github.event_name == 'workflow_dispatch'
-    runs-on: ubuntu-latest
-    environment: production
-    steps:
-      - uses: actions/checkout@v4${config.deploy === 'kustomize' ? `
-      - name: Apply Kustomize
-        run: |
-          kustomize build kustomize/overlays/prod > /tmp/manifests.yaml
-          kubectl apply --dry-run=server -f /tmp/manifests.yaml
-          kubectl apply -f /tmp/manifests.yaml` : config.deploy === 'helm' ? `
-      - name: Helm upgrade
-        run: |
-          helm upgrade --install threshold-exporter \\
-            oci://ghcr.io/vencil/charts/threshold-exporter \\
-            -f environments/prod/values.yaml \\
-            -n monitoring --wait` : `
-      - name: Trigger ArgoCD sync
-        run: argocd app sync dynamic-alerting --force`}`;
-}
+const generateInitCommand = window.__cicdGenerateInitCommand;
+const generateDockerCommand = window.__cicdGenerateDockerCommand;
+const generateFileTree = window.__cicdGenerateFileTree;
+const generateGitHubActionsPreview = window.__cicdGenerateGitHubActionsPreview;
 
 /* ── Step components ── */
