+++ docs/interactive/tools/deployment-wizard.jsx
@@ -5,456 +5,35 @@ audience: [platform-engineer, sre, devops]
 version: v2.7.0
 lang: en
 related: [architecture-quiz, cicd-setup-wizard, capacity-planner]
+dependencies: [
+  "deployment-wizard/fixtures/wizard-defaults.js",
+  "deployment-wizard/utils/generators.js"
+]
 ---
 
 /* eslint-disable */
+// PR-portal-10: data + helpers extracted to sibling subdirectory
+// (mirrors operator-setup-wizard PR-portal-4 + cicd-setup-wizard PR-10).
+// Step components remain inline; future PR can extract them.
+//
 // Design token migration notes (v2.7.0 Phase .a0, DEC-F/G/I):
-//   • DEC-F (C): removed all Tailwind `dark:` variants; rely on `[data-theme="dark"]`
+//   * DEC-F (C): removed all Tailwind `dark:` variants; rely on `[data-theme="dark"]`
 //     attribute + design-tokens.css to swap `--da-color-*` automatically.
-//   • DEC-G (B): Portal-wide slate→gray unification is implicit here — all slate/blue
-//     shade classes were migrated to `var(--da-color-*)` tokens, so no literal slate/gray
-//     remains except the fixed IDE-style dark code-preview panel (line ~820).
-//   • REG-004: hardcoded portal paths (/template-gallery, /tenant-manager,
-//     /docs/getting-started) replaced with proper `../../assets/jsx-loader.html?component=`
-//     and relative docs paths so they resolve inside the jsx-loader portal.
-//   • Residuals: `bg-slate-900 text-slate-100` in Review step code preview (IDE-intent),
-//     `text-amber-800` for warning message copy (tokens don't include warning-fg variant).
-//   • Follow-ups (v2.8.0): jsx-loader `navigate(key)` helper (DEC-I C) will replace the
-//     raw href strings; add `--da-color-warning-fg` token to finish amber migration.
+//   * DEC-G (B): Portal-wide slate->gray unification is implicit here.
+//   * REG-004: hardcoded portal paths replaced with proper jsx-loader URLs.
 import React, { useState, useMemo } from 'react';
 
 const t = window.__t || ((zh, en) => en);
 
-/* ── Step definitions ── */
-const STEPS = [
-  { id: 'tier', label: () => t('部署層級', 'Deployment Tier') },
-  { id: 'environment', label: () => t('運行環境', 'Environment') },
-  { id: 'tenants', label: () => t('Tenant 數量', 'Tenant Count') },
-  { id: 'auth', label: () => t('認證 (Tier 2)', 'Authentication (Tier 2)') },
-  { id: 'packs', label: () => t('Rule Packs', 'Rule Packs') },
-  { id: 'review', label: () => t('檢視與產出', 'Review & Generate') },
-];
+const STEPS = window.__DEPLOY_STEPS;
+const TIERS = window.__DEPLOY_TIERS;
+const ENVIRONMENTS = window.__DEPLOY_ENVIRONMENTS;
+const TENANT_SIZES = window.__DEPLOY_TENANT_SIZES;
+const OAUTH2_PROVIDERS = window.__DEPLOY_OAUTH2_PROVIDERS;
+const RULE_PACKS = window.__DEPLOY_RULE_PACKS;
 
-/* ── Deployment tiers ── */
-const TIERS = [
-  {
-    id: 'tier1',
-    name: t('Tier 1：Git-Native', 'Tier 1: Git-Native'),
-    desc: t('純 GitOps：YAML + da-tools CLI + Helm values', 'Pure GitOps: YAML + da-tools CLI + Helm values'),
-    features: [
-      t('threshold-exporter × 2 (HA)', 'threshold-exporter × 2 (HA)'),
-      t('Prometheus + Alertmanager (Helm)', 'Prometheus + Alertmanager (Helm)'),
-      t('ConfigMap 管理告警規則', 'ConfigMap for alert rules'),
-      t('無 Portal / API', 'No Portal / API'),
-    ],
-    icon: '📦',
-    cost: t('低', 'Low'),
-  },
-  {
-    id: 'tier2',
-    name: t('Tier 2：Portal + API', 'Tier 2: Portal + API'),
-    desc: t('完整功能：Tier 1 + da-portal + tenant-api + OAuth2', 'Full-featured: Tier 1 + da-portal + tenant-api + OAuth2'),
-    features: [
-      t('所有 Tier 1 功能', 'All Tier 1 features'),
-      t('da-portal UI (自託管或 SaaS)', 'da-portal UI (self-hosted or SaaS)'),
-      t('tenant-api（RBAC + 熱更新）', 'tenant-api (RBAC + hot-reload)'),
-      t('oauth2-proxy（GitHub / Google / OIDC）', 'oauth2-proxy (GitHub / Google / OIDC)'),
-    ],
-    icon: '🌐',
-    cost: t('中', 'Medium'),
-  },
-];
+const generateHelmValues = window.__deployGenerateHelmValues;
 
-/* ── Environments ── */
-const ENVIRONMENTS = [
-  {
-    id: 'local',
-    label: t('本地開發 (Kind/Minikube)', 'Local Dev (Kind/Minikube)'),
-    icon: '💻',
-    desc: t('2–4 CPU, 4–8 GB RAM, 簡化部署', '2–4 CPU, 4–8 GB RAM, simplified'),
-  },
-  {
-    id: 'staging',
-    label: t('測試環境 (Staging)', 'Staging Environment'),
-    icon: '🧪',
-    desc: t('4–8 CPU, 16 GB RAM, HA 就緒', '4–8 CPU, 16 GB RAM, HA-ready'),
-  },
-  {
-    id: 'production',
-    label: t('生產環境 (Production)', 'Production Environment'),
-    icon: '🚀',
-    desc: t('8+ CPU, 32+ GB RAM, 多區域', '8+ CPU, 32+ GB RAM, multi-region'),
-  },
-];
-
-/* ── Tenant sizes ── */
-const TENANT_SIZES = [
-  {
-    id: 'small',
-    label: t('小型 (1–10)', 'Small (1–10)'),
-    icon: '1️⃣',
-    replicas: { exporter: 1, prometheus: 1, alertmanager: 1 },
-    retention: '7d',
-    cardinality: 500,
-  },
-  {
-    id: 'medium',
-    label: t('中型 (10–50)', 'Medium (10–50)'),
-    icon: '📊',
-    replicas: { exporter: 2, prometheus: 2, alertmanager: 3 },
-    retention: '14d',
-    cardinality: 2000,
-  },
-  {
-    id: 'large',
-    label: t('大型 (50+)', 'Large (50+)'),
-    icon: '📈',
-    replicas: { exporter: 3, prometheus: 3, alertmanager: 3 },
-    retention: '30d',
-    cardinality: 5000,
-  },
-];
-
-/* ── OAuth2 providers ── */
-const OAUTH2_PROVIDERS = [
-  {
-    id: 'github',
-    label: 'GitHub',
-    icon: '🐙',
-    desc: t('使用 GitHub 帳戶登入', 'Sign in with GitHub account'),
-    scopes: ['user:email', 'read:org'],
-  },
-  {
-    id: 'google',
-    label: 'Google',
-    icon: '🔵',
-    desc: t('使用 Google 帳戶登入', 'Sign in with Google account'),
-    scopes: ['openid', 'email', 'profile'],
-  },
-  {
-    id: 'oidc',
-    label: 'OIDC / Keycloak',
-    icon: '🔐',
-    desc: t('自託管 OIDC（Keycloak、Okta 等）', 'Self-hosted OIDC (Keycloak, Okta, etc.)'),
-    scopes: ['openid', 'profile', 'email'],
-  },
-  {
-    id: 'gitlab',
-    label: 'GitLab',
-    icon: '🦊',
-    desc: t('使用 GitLab 帳戶登入', 'Sign in with GitLab account'),
-    scopes: ['openid', 'profile', 'email'],
-  },
-];
-
-/* ── Rule packs ── */
-const RULE_PACKS = [
-  { id: 'mariadb', label: 'MariaDB/MySQL', category: 'database', icon: '🐬' },
-  { id: 'postgresql', label: 'PostgreSQL', category: 'database', icon: '🐘' },
-  { id: 'redis', label: 'Redis', category: 'database', icon: '🔴' },
-  { id: 'mongodb', label: 'MongoDB', category: 'database', icon: '🍃' },
-  { id: 'elasticsearch', label: 'Elasticsearch', category: 'database', icon: '🔎' },
-  { id: 'oracle', label: 'Oracle', category: 'database', icon: '🏛️' },
-  { id: 'db2', label: 'DB2', category: 'database', icon: '🔷' },
-  { id: 'clickhouse', label: 'ClickHouse', category: 'database', icon: '🖱️' },
-  { id: 'kafka', label: 'Kafka', category: 'messaging', icon: '📨' },
-  { id: 'rabbitmq', label: 'RabbitMQ', category: 'messaging', icon: '🐰' },
-  { id: 'jvm', label: 'JVM', category: 'runtime', icon: '☕' },
-  { id: 'nginx', label: 'Nginx', category: 'webserver', icon: '🌐' },
-  { id: 'kubernetes', label: 'Kubernetes', category: 'infrastructure', icon: '⎈' },
-];
-
-/* ── Helm values generator ── */
-function generateHelmValues(config) {
-  const { tier, environment, tenantSize, auth, packs } = config;
-  const size = TENANT_SIZES.find(s => s.id === tenantSize);
-  const isTier2 = tier === 'tier2';
-
-  let yaml = `# Generated Helm values for ${TIERS.find(t => t.id === tier)?.name}
-# Environment: ${ENVIRONMENTS.find(e => e.id === environment)?.label}
-# Tenant count: ${size?.label}
-# Generated: ${new Date().toISOString().split('T')[0]}
-
-# ────────────────────────────────────────────────────────────────────
-# threshold-exporter Configuration
-# ────────────────────────────────────────────────────────────────────
-
-thresholdExporter:
-  replicaCount: ${size?.replicas.exporter || 2}
-  image:
-    repository: ghcr.io/vencil/threshold-exporter
-    tag: v2.7.0
-    pullPolicy: IfNotPresent
-
-  resources:
-    requests:
-      cpu: ${environment === 'local' ? '100m' : environment === 'staging' ? '250m' : '500m'}
-      memory: ${environment === 'local' ? '128Mi' : environment === 'staging' ? '256Mi' : '512Mi'}
-    limits:
-      cpu: ${environment === 'local' ? '200m' : environment === 'staging' ? '500m' : '1000m'}
-      memory: ${environment === 'local' ? '256Mi' : environment === 'staging' ? '512Mi' : '1Gi'}
-
-  # Hot-reload SHA-256 validation
-  configValidation:
-    enabled: true
-    sha256: "" # Set after generating config
-
-  # Cardinality guard: per-tenant max metrics
-  cardinalityGuard:
-    enabled: true
-    maxPerTenant: ${size?.cardinality || 2000}
-
-  # Three-state operating modes: normal / silent / maintenance
-  tripleState:
-    enabled: true
-    defaultMode: normal
-
-prometheus:
-  replicaCount: ${size?.replicas.prometheus || 2}
-  image:
-    repository: prom/prometheus
-    tag: v2.52.0
-
-  resources:
-    requests:
-      cpu: ${environment === 'local' ? '250m' : environment === 'staging' ? '500m' : '1000m'}
-      memory: ${environment === 'local' ? '512Mi' : environment === 'staging' ? '1Gi' : '2Gi'}
-    limits:
-      cpu: ${environment === 'local' ? '500m' : environment === 'staging' ? '1000m' : '2000m'}
-      memory: ${environment === 'local' ? '1Gi' : environment === 'staging' ? '2Gi' : '4Gi'}
-
-  # Data retention based on tenant size
-  retention: "${size?.retention || '14d'}"
-
-  # Rule packs from ConfigMap + Projected Volume
-  ruleConfigMaps:
-    - name: platform-rules
-      key: rules.yaml
-    ${packs.length > 0 ? `# Auto-mounted rule packs via Projected Volume:\n    # ${packs.map(p => `- name: rules-${p}`).join('\n    # ')}` : ''}
-
-  # ServiceMonitor for threshold-exporter
-  serviceMonitor:
-    enabled: true
-    interval: 30s
-    scrapeTimeout: 10s
-
-alertmanager:
-  replicaCount: ${size?.replicas.alertmanager || 3}
-  image:
-    repository: prom/alertmanager
-    tag: v0.27.0
-
-  resources:
-    requests:
-      cpu: ${environment === 'local' ? '100m' : environment === 'staging' ? '250m' : '500m'}
-      memory: ${environment === 'local' ? '128Mi' : environment === 'staging' ? '256Mi' : '512Mi'}
-    limits:
-      cpu: ${environment === 'local' ? '200m' : environment === 'staging' ? '500m' : '1000m'}
-      memory: ${environment === 'local' ? '256Mi' : environment === 'staging' ? '512Mi' : '1Gi'}
-
-  # Dynamic route generation + configmap-reload
-  configReload:
-    enabled: true
-    image: jimmidyson/configmap-reload:v0.5.0
-
-  # Cluster mode for HA
-  clustering:
-    enabled: ${environment !== 'local' ? 'true' : 'false'}
-    peers:
-      enabled: ${environment !== 'local' ? 'true' : 'false'}
-
-# ────────────────────────────────────────────────────────────────────
-# Platform Common Settings
-# ────────────────────────────────────────────────────────────────────
-
-platform:
-  # Environment label for metric routing
-  environment: ${environment}
-
-  # Namespace isolation
-  namespaces:
-    monitoring: monitoring
-    # Add tenant namespaces as needed
-
-  # Logging level
-  logLevel: ${environment === 'production' ? 'warn' : 'info'}
-
-  # Bilingual support (zh/en annotations)
-  i18n:
-    enabled: true
-    defaultLanguage: en
-
-# ────────────────────────────────────────────────────────────────────
-# Tier 2: Portal + API Configuration
-# ────────────────────────────────────────────────────────────────────
-${isTier2 ? `
-daPortal:
-  enabled: true
-  replicaCount: ${environment === 'local' ? 1 : size?.replicas.exporter || 2}
-  image:
-    repository: ghcr.io/vencil/da-portal
-    tag: v2.7.0
-
-  resources:
-    requests:
-      cpu: ${environment === 'local' ? '100m' : '250m'}
-      memory: ${environment === 'local' ? '256Mi' : '512Mi'}
-    limits:
-      cpu: ${environment === 'local' ? '200m' : '500m'}
-      memory: ${environment === 'local' ? '512Mi' : '1Gi'}
-
-  # Portal ingress
-  ingress:
-    enabled: true
-    className: nginx
-    hosts:
-      - host: da-portal.example.com
-        paths:
-          - path: /
-            pathType: Prefix
-
-tenantAPI:
-  enabled: true
-  replicaCount: ${environment === 'local' ? 1 : 2}
-  image:
-    repository: ghcr.io/vencil/tenant-api
-    tag: v2.7.0
-
-  resources:
-    requests:
-      cpu: ${environment === 'local' ? '100m' : '250m'}
-      memory: ${environment === 'local' ? '128Mi' : '256Mi'}
-    limits:
-      cpu: ${environment === 'local' ? '200m' : '500m'}
-      memory: ${environment === 'local' ? '256Mi' : '512Mi'}
-
-  # RBAC hot-reload via atomic.Value
-  rbac:
-    enabled: true
-    cacheRefreshInterval: 30s
-
-  # Tenant API ingress
-  ingress:
-    enabled: true
-    className: nginx
-    hosts:
-      - host: api.dynamic-alerting.example.com
-        paths:
-          - path: /v1
-            pathType: Prefix
-
-oauth2Proxy:
-  enabled: true
-  replicaCount: ${environment === 'local' ? 1 : 2}
-  image:
-    repository: oauth2-proxy/oauth2-proxy
-    tag: v7.6.0
-
-  resources:
-    requests:
-      cpu: 50m
-      memory: 64Mi
-    limits:
-      cpu: 100m
-      memory: 128Mi
-
-  # OAuth2 provider configuration
-  config:
-    provider: "${auth || 'oidc'}"
-    ${auth === 'github' ? `oauth_url: "https://github.com/login/oauth/authorize"
-    token_url: "https://github.com/login/oauth/access_token"
-    user_info_url: "https://api.github.com/user"
-    scopes: ["user:email", "read:org"]` : auth === 'google' ? `oauth_url: "https://accounts.google.com/o/oauth2/v2/auth"
-    token_url: "https://oauth2.googleapis.com/token"
-    user_info_url: "https://www.googleapis.com/oauth2/v2/userinfo"
-    scopes: ["openid", "email", "profile"]` : auth === 'gitlab' ? `oauth_url: "https://gitlab.com/oauth/authorize"
-    token_url: "https://gitlab.com/oauth/token"
-    user_info_url: "https://gitlab.com/api/v4/user"
-    scopes: ["openid", "profile", "email"]` : `oauth_url: "https://your-keycloak.com/auth/realms/master/protocol/openid-connect/auth"
-    token_url: "https://your-keycloak.com/auth/realms/master/protocol/openid-connect/token"
-    user_info_url: "https://your-keycloak.com/auth/realms/master/protocol/openid-connect/userinfo"
-    scopes: ["openid", "profile", "email"]`}
-    client_id: "" # Set in secrets
-    client_secret: "" # Set in secrets
-
-  # Cookie configuration for session persistence
-  cookie:
-    domain: example.com
-    secure: true
-    httponly: true
-    samesite: Lax
-` : `
-# Tier 1: Portal and API disabled
-daPortal:
-  enabled: false
-
-tenantAPI:
-  enabled: false
-
-oauth2Proxy:
-  enabled: false
-`}
-
-# ────────────────────────────────────────────────────────────────────
-# Networking & Storage
-# ────────────────────────────────────────────────────────────────────
-
-persistence:
-  # Prometheus TSDB storage
-  prometheus:
-    enabled: true
-    storageClass: standard
-    size: ${environment === 'local' ? '5Gi' : environment === 'staging' ? '20Gi' : '100Gi'}
-
-  # Alertmanager state
-  alertmanager:
-    enabled: true
-    storageClass: standard
-    size: ${environment === 'local' ? '1Gi' : '5Gi'}
-
-networkPolicy:
-  enabled: ${environment === 'production' ? 'true' : 'false'}
-  ingressNamespaces:
-    - monitoring
-
-# ────────────────────────────────────────────────────────────────────
-# Observability & Debugging
-# ────────────────────────────────────────────────────────────────────
-
-monitoring:
-  # Prometheus scrape config for self-monitoring
-  prometheus:
-    enabled: true
-    interval: 60s
-
-  # Log aggregation hints
-  logging:
-    level: ${environment === 'production' ? 'warn' : 'info'}
-    format: json
-
-# ────────────────────────────────────────────────────────────────────
-# Security
-# ────────────────────────────────────────────────────────────────────
-
-rbac:
-  create: true
-
-serviceAccount:
-  create: true
-  name: threshold-exporter
-
-podSecurityPolicy:
-  enabled: ${environment === 'production' ? 'true' : 'false'}
-
-# Secrets for OAuth2 (if Tier 2)
-secrets:
-  ${isTier2 ? `oauth2:
-    clientId: "" # Fill from secrets manager
-    clientSecret: "" # Fill from secrets manager
-  ` : ''}# Add any additional secrets here
-`;
-  return yaml;
-}
 
 export default function DeploymentWizard() {
   const [step, setStep] = useState(0);
